Updated UnitTypes::topSpeed so that it correctly returns the unit's top speed, in pixels per frame. The return type has been changed to double.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@1560 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/include/BWAPI/UnitType.h
@@ -62,7 +62,7 @@ namespace BWAPI
       const WeaponType* airWeapon() const;
       int maxAirHits() const;
 
-      int topSpeed() const;
+      double topSpeed() const;
       int acceleration() const;
       int haltDistance() const;
       int turnRadius() const;
